+++ src/Core/Serialization/JSON/Normalizers/ExternalReferenceRepositoryNormalizer.php
@@ -25,6 +25,8 @@ namespace CycloneDX\Core\Serialization\JSON\Normalizers;
 
 use CycloneDX\Core\Collections\ExternalReferenceRepository;
 use CycloneDX\Core\Serialization\JSON\_BaseNormalizer;
+use DomainException;
+use UnexpectedValueException;
 
 /**
  * @author jkowalleck
@@ -44,7 +46,7 @@ class ExternalReferenceRepositoryNormalizer extends _BaseNormalizer
         foreach ($repo->getItems() as $externalReference) {
             try {
                 $item = $normalizer->normalize($externalReference);
-            } catch (\DomainException) {
+            } catch (DomainException|UnexpectedValueException) {
                 continue;
             }
             if (false === empty($item)) {